+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Version TBD
+_Date TBD_
+
+### Changes
+* Find any current products that have bad values for their criterions' Privacy and Security Framework
+
 ## Version 8.5.0
 _24 April 2017_